Added discount mechanism
Stored procedures `checkin` and `checkout` both modify the number of
discount tickets per station

+++ src/benchmarks/edu/brown/benchmark/bikerstream/bikerstream-ddl.sql
@@ -17,7 +17,7 @@ CREATE TABLE StationStatus
     station_id            INTEGER NOT NULL REFERENCES stations(station_id)
 ,   current_bikes         INTEGER NOT NULL
 ,   current_docks         INTEGER NOT NULL
-,   current_discount      FLOAT   NOT NULL
+,   current_discount      INTEGER NOT NULL
 );
 
 -- Keep track of riders in the system.
